feat: moving canvas controls to their own file
