+++ client/src/components/Canvas/Canvas.js
@@ -1,15 +1,7 @@
 import React from "react";
 import { connect } from "react-redux";
 import { fabric } from "fabric";
-import { saveAs } from "file-saver";
-import {
-  Button,
-  Dropdown,
-  DropdownButton,
-  Tooltip,
-  OverlayTrigger,
-  Container,
-} from "react-bootstrap";
+import { Button, Dropdown, DropdownButton, Container } from "react-bootstrap";
 import "../../styles/canvas.css";
 import { fourPanel, threePanel, sixPanel, removePanel } from "./Templates";
 import { AddTextBox } from "./AddTextBox";
@@ -19,7 +11,7 @@ import Bubbles from "../TextBubbles/Bubbles";
 import Characters from "../Characters/characters";
 import { fetchCanvasElements, saveCanvasElements } from "../../store/index";
 import ColorPicker from "../Editor/ColorPicker";
-import { canvasControlsCopy } from "./Copy";
+import CanvasControls from "./CanvasControls";
 
 let windowHeightRatio = Math.floor(0.7 * window.innerHeight);
 let windowWidthRatio = Math.floor(0.7 * window.innerWidth);
@@ -31,14 +23,7 @@ class Canvas extends React.Component {
       canvas: {},
       selectedCanvasId: "canvas",
     };
-
     this.initCanvas = this.initCanvas.bind(this);
-    this.removeObject = this.removeObject.bind(this);
-    this.save = this.save.bind(this);
-    this.sendFront = this.sendFront.bind(this);
-    this.sendBack = this.sendBack.bind(this);
-    this.saveToStore = this.saveToStore.bind(this);
-
     this.createEventListener = this.createEventListener.bind(this);
     this.deleteWithKeyboard = this.deleteWithKeyboard.bind(this);
   }
@@ -78,10 +63,6 @@ class Canvas extends React.Component {
     }
   }
 
-  saveToStore = (canvas, selectedCanvasId) => {
-    this.props.saveCanvas(canvas.getObjects(), selectedCanvasId);
-  };
-
   initCanvas = () =>
     new fabric.Canvas("canvas", {
       //1:1 ratio
@@ -89,67 +70,6 @@ class Canvas extends React.Component {
       width: windowWidthRatio,
       backgroundColor: "white",
     });
-
-  // crossOrigin = anonymous before save needed
-  save = () => {
-    var canvas = document.getElementById("canvas");
-    canvas.toBlob(function (blob) {
-      // let downloadedImg = new Image(blob);
-      // downloadedImg.crossOrigin = "Anonymous";
-      // blob.crossOrigin = "Anonymous";
-      saveAs(blob, "comic.png");
-      // saveAs(downloadedImg, 'comic.png');
-    });
-  };
-
-  removeObject = (canvas) => {
-    let activeObject = canvas.getActiveObjects();
-    if (activeObject) {
-      canvas.discardActiveObject();
-      activeObject.forEach(function (object) {
-        canvas.remove(object);
-      });
-    }
-  };
-
-  clearCanvas = (canvas) => {
-    canvas.getObjects().forEach((obj) => {
-      canvas.getActiveObject(obj);
-      canvas.remove(obj);
-    });
-  };
-
-  sendFrontOne = (canvas) => {
-    const activeObject = canvas.getActiveObjects();
-    activeObject.forEach((object) => {
-      object.bringForward();
-      canvas.renderAll();
-    });
-  };
-
-  sendFront = (canvas) => {
-    const activeObject = canvas.getActiveObjects();
-    activeObject.forEach((object) => {
-      object.bringToFront();
-      canvas.renderAll();
-    });
-  };
-
-  sendBackOne = (canvas) => {
-    const activeObject = canvas.getActiveObjects();
-    activeObject.forEach((object) => {
-      object.sendBackwards();
-      canvas.renderAll();
-    });
-  };
-
-  sendBack = (canvas) => {
-    const activeObject = canvas.getActiveObjects();
-    activeObject.forEach((object) => {
-      object.sendToBack();
-      canvas.renderAll();
-    });
-  };
 
   createEventListener() {
     document.addEventListener("keydown", this.deleteWithKeyboard);
@@ -217,111 +137,11 @@ class Canvas extends React.Component {
 
               <Characters canvasInstance={canvasInstance} />
               <Bubbles canvasInstance={canvasInstance} />
-              {/* send up just one layer */}
-              <OverlayTrigger
-                placement="top"
-                overlay={<Tooltip>{canvasControlsCopy.bringUpOne}</Tooltip>}
-              >
-                <Button
-                  variant="light"
-                  onClick={() => this.sendFrontOne(canvasInstance)}
-                >
-                  <i className="fas fa-angle-up"></i>
-                </Button>
-              </OverlayTrigger>
 
-              {/* send all the way to top layer */}
-              <OverlayTrigger
-                placement="top"
-                overlay={<Tooltip>{canvasControlsCopy.bringUp}</Tooltip>}
-              >
-                <Button
-                  variant="light"
-                  onClick={() => this.sendFront(canvasInstance)}
-                >
-                  <i className="fas fa-angle-double-up"></i>
-                </Button>
-              </OverlayTrigger>
-
-              {/* send down just one layer */}
-              <OverlayTrigger
-                placement="top"
-                overlay={<Tooltip>{canvasControlsCopy.bringDownOne}</Tooltip>}
-              >
-                <Button
-                  variant="light"
-                  onClick={() => this.sendBackOne(canvasInstance)}
-                >
-                  <i className="fas fa-angle-down"></i>
-                </Button>
-              </OverlayTrigger>
-
-              {/* send all the way to bottom layer */}
-              <OverlayTrigger
-                placement="top"
-                overlay={<Tooltip>{canvasControlsCopy.bringDown}</Tooltip>}
-              >
-                <Button
-                  variant="light"
-                  onClick={() => this.sendBack(canvasInstance)}
-                >
-                  <i className="fas fa-angle-double-down"></i>
-                </Button>
-              </OverlayTrigger>
-
-              {/* save to store button */}
-              <OverlayTrigger
-                placement="top"
-                overlay={<Tooltip>{canvasControlsCopy.save}</Tooltip>}
-              >
-                <Button
-                  variant="light"
-                  onClick={() =>
-                    this.saveToStore(
-                      canvasInstance,
-                      this.state.selectedCanvasId
-                    )
-                  }
-                >
-                  <i className="far fa-save"></i>
-                </Button>
-              </OverlayTrigger>
-
-              {/* download as image button */}
-              <OverlayTrigger
-                placement="top"
-                overlay={<Tooltip>{canvasControlsCopy.download}</Tooltip>}
-              >
-                <Button variant="light" onClick={() => this.save()}>
-                  <i className="fas fa-file-download"></i>
-                </Button>
-              </OverlayTrigger>
-
-              {/* delete selected element(s) button */}
-              <OverlayTrigger
-                placement="top"
-                overlay={<Tooltip>{canvasControlsCopy.delete}</Tooltip>}
-              >
-                <Button
-                  variant="light"
-                  onClick={() => this.removeObject(canvasInstance)}
-                >
-                  <i className="fas fa-eraser"></i>
-                </Button>
-              </OverlayTrigger>
-
-              {/* clear canvas button */}
-              <OverlayTrigger
-                placement="top"
-                overlay={<Tooltip>{canvasControlsCopy.clearCanvas}</Tooltip>}
-              >
-                <Button
-                  variant="outline-danger"
-                  onClick={() => this.clearCanvas(canvasInstance)}
-                >
-                  <i className="far fa-trash-alt"></i>
-                </Button>
-              </OverlayTrigger>
+              <CanvasControls
+                canvasInstance={canvasInstance}
+                selectedCanvasId={this.state.selectedCanvasId}
+              />
             </Container>
             <canvas id={`canvas`} width="300" height="300" />
           </div>
